+++ app.js
@@ -3,6 +3,7 @@
 'use strict';
 
 const clc             = require('cli-color');
+const csrf            = require('csurf');
 const commander       = require('commander');
 const express         = require('express');
 const fs              = require('fs');
@@ -13,21 +14,18 @@ const updateNotifier  = require('update-notifier');
 const pkg             = require('./package.json');
 
 let app               = express();
-let notifier          = updateNotifier({pkg});
+let notifier          = updateNotifier({ pkg });
 
 let config;
 let defaultPort = 80;
 let server      = app;
 let sslOptions;
 
-console.log('Welcome to mongo-express');
-console.log('------------------------');
-console.log('\n');
-
 // Notify of any updates
 notifier.notify();
 
 try {
+  // eslint-disable-next-line import/no-unresolved
   config = utils.deepmerge(require('./config.default'), require('./config'));
 } catch (e) {
   if (e.code === 'MODULE_NOT_FOUND') {
@@ -42,8 +40,16 @@ try {
   config = require('./config.default');
 }
 
+if (config.options.console) {
+  console.log('Welcome to mongo-express');
+  console.log('------------------------');
+  console.log('\n');
+}
+
 commander
   .version(require('./package').version)
+  .option('-H, --host <host>', 'hostname or adress')
+  .option('-P, --dbport <host>', 'port of the db')
   .option('-u, --username <username>', 'username for authentication')
   .option('-p, --password <password>', 'password for authentication')
   .option('-a, --admin', 'enable authentication as admin')
@@ -74,6 +80,9 @@ if (commander.username && commander.password) {
   config.useBasicAuth = false;
 }
 
+config.mongodb.server = commander.host || config.mongodb.server;
+config.mongodb.port = commander.dbport || config.mongodb.port;
+
 config.site.port = commander.port || config.site.port;
 
 if (!config.site.baseUrl) {
@@ -81,17 +90,8 @@ if (!config.site.baseUrl) {
   config.site.baseUrl = '/';
 }
 
-if (config.basicAuth.username === 'admin' && config.basicAuth.password === 'pass') {
-  console.error(clc.red('basicAuth credentials are "admin:pass", it is recommended you change this in your config.js!'));
-}
-
-if (!config.site.host || config.site.host === '0.0.0.0') {
-  console.error(clc.red('Server is open to allow connections from anyone (0.0.0.0)'));
-}
-
 app.use(config.site.baseUrl, middleware(config));
-app.set('read_only',      config.options.readOnly       || false);
-app.set('gridFSEnabled',  config.options.gridFSEnabled  || false);
+app.use(csrf());
 
 if (config.site.sslEnabled) {
   defaultPort     = 443;
@@ -102,8 +102,32 @@ if (config.site.sslEnabled) {
   server = https.createServer(sslOptions, app);
 }
 
-server.listen(config.site.port, config.site.host, function() {
-  console.log('Mongo Express server listening',
-    'on port ' + (config.site.port || defaultPort),
-    'at '      + (config.site.host || '0.0.0.0'));
+let addressString = (config.site.sslEnabled ? 'https://' : 'http://') + (config.site.host || '0.0.0.0') + ':' + (config.site.port || defaultPort);
+
+server.listen(config.site.port, config.site.host, function () {
+  if (config.options.console) {
+
+    console.log('Mongo Express server listening', 'at ' + addressString);
+
+    if (!config.site.host || config.site.host === '0.0.0.0') {
+      console.error(clc.red('Server is open to allow connections from anyone (0.0.0.0)'));
+    }
+
+    if (config.basicAuth.username === 'admin' && config.basicAuth.password === 'pass') {
+      console.error(clc.red('basicAuth credentials are "admin:pass", it is recommended you change this in your config.js!'));
+    }
+
+  }
+})
+.on('error', function (e) {
+  if (e.code === 'EADDRINUSE') {
+    console.log();
+    console.error(clc.red('Address ' + addressString + ' already in use! You need to pick a different host and/or port.'));
+    console.log('Maybe mongo-express is already running?');
+  }
+
+  console.log();
+  console.log('If you are still having trouble, try Googling for the key parts of the following error object before posting an issue');
+  console.log(JSON.stringify(e));
+  return process.exit(1);
 });